refactor(migration): extract db_utils.py to share SQL connection helpers (#1332)
Eliminates duplicate connection logic between migrate-fake-groups.py
and the test conftest.
