+++ scripts/migrate-fake-groups.py
@@ -21,102 +21,15 @@ Requires:
 
 import argparse
 import json
-import subprocess
 import sys
 import uuid
 from datetime import datetime, timezone
 from pathlib import Path
 
-try:
-    import pyodbc
-except ImportError:
-    print("ERROR: pyodbc not installed. Run: pip install pyodbc")
-    sys.exit(1)
+from db_utils import connect_azure, connect_local
 
 SCRIPTS_DIR = Path(__file__).resolve().parent
 ALLOWLIST_PATH = SCRIPTS_DIR / "fake-groups-allowlist.json"
-
-KEY_VAULT_NAME = "kv-lt-dev-5ba22u"
-KEY_VAULT_SECRET = "ConnectionStrings--Default"
-
-LOCAL_SERVER = "tcp:127.0.0.1,1434"
-LOCAL_DATABASE = "LangTeach"
-
-
-# ---------------------------------------------------------------------------
-# Connection helpers (copied from copy-azure-teacher.py)
-# ---------------------------------------------------------------------------
-
-def get_azure_connection_string():
-    result = subprocess.run(
-        ["az", "keyvault", "secret", "show",
-         "--vault-name", KEY_VAULT_NAME, "--name", KEY_VAULT_SECRET,
-         "--query", "value", "-o", "tsv"],
-        capture_output=True, text=True
-    )
-    if result.returncode != 0:
-        print(f"ERROR: Failed to read Key Vault secret. Run 'az login' first.\n{result.stderr}")
-        sys.exit(1)
-    return result.stdout.strip()
-
-
-def parse_sqlserver_conn_string(conn_str):
-    parts = {}
-    for segment in conn_str.split(";"):
-        segment = segment.strip()
-        if "=" in segment:
-            key, value = segment.split("=", 1)
-            parts[key.strip().lower()] = value.strip()
-    return parts
-
-
-def connect_azure():
-    raw_conn = get_azure_connection_string()
-    parts = parse_sqlserver_conn_string(raw_conn)
-
-    server = parts.get("server", "")
-    database = parts.get("initial catalog") or parts.get("database", "")
-    user_id = parts.get("user id") or parts.get("uid", "")
-    password = parts.get("password", "")
-
-    conn_str = (
-        f"DRIVER={{ODBC Driver 18 for SQL Server}};"
-        f"SERVER={server};"
-        f"DATABASE={database};"
-        f"UID={user_id};"
-        f"PWD={password};"
-        f"Encrypt=yes;"
-        f"TrustServerCertificate=no;"
-        f"Connection Timeout=30;"
-    )
-    conn = pyodbc.connect(conn_str)
-    conn.autocommit = False
-    return conn
-
-
-def connect_local():
-    env_path = SCRIPTS_DIR.parent / ".env"
-    sa_password = None
-    if env_path.exists():
-        for line in env_path.read_text().splitlines():
-            if line.startswith("SA_PASSWORD="):
-                sa_password = line.split("=", 1)[1].strip()
-                break
-    if not sa_password:
-        print("ERROR: SA_PASSWORD not found in .env")
-        sys.exit(1)
-
-    conn_str = (
-        f"DRIVER={{ODBC Driver 18 for SQL Server}};"
-        f"SERVER={LOCAL_SERVER};"
-        f"DATABASE={LOCAL_DATABASE};"
-        f"UID=sa;"
-        f"PWD={sa_password};"
-        f"TrustServerCertificate=yes;"
-    )
-    conn = pyodbc.connect(conn_str)
-    conn.autocommit = False
-    return conn
 
 
 # ---------------------------------------------------------------------------
